requirements.txt
# Core dependencies
python-dotenv>=1.0.0
asyncio
argparse

# Agent dependencies
crewai>=0.1.0
crewai-tools>=0.1.0
langchain-openai
playwright>=1.40.0
httpx
beautifulsoup4>=4.12.0
aider-chat>=0.1.0
pyyaml>=6.0
pydantic>=2.0.0

# Performance and analysis
litellm>=1.0.0
tiktoken>=0.5.0
undici

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 